Eliminated StorageClasses, replaced MapStore class with map
The MapStore class was a temporary dictionary-like class created for the
early prototype. This version replaces it with the std::map class. Note
that the std::map class uses iterators which complicated the code
slightly because the MapStore class was iterable by index values. This
made it easier to create an array of indexable files. For this simple
version, a new map was created to translate the object name (key) to
an integer reference into the file array. For some reason, compile-time
errors were generated when trying to make a map of <string,FILE**>
directly.

This commit also moves the StorageClasses files into a "storage"
directory in the "experiments" directory (git mv). This was done to keep
the main directory relatively clean. It already contains several
prototype versions (C++, C, Python).