Added toy_scheduler.cpp to experiment witn MCell's current scheduler.
This is currently a stand-alone C++ program that incorporates some of
the code from sched_util.c converted to C++ classes. The main program
creates a baseline timestep scheduling window and then inserts events at
varying time scales (from 0.3 up to 5000000.0) to observe the behavior.

This toy scheduler does not depend on any other code and is compiled
with a simple g++ command (not in the make file).