Added an ability to delete by index number rather than address.
The scheduler deletion uses an address of the item being deleted. This
is inconvenient to provide via a text interface, so this code maps these
addresses to integers to allow easy deletion. As with all of the
scheduler testing code, this didn't require any changes to the
scheduling code itself.