Replace C version of embedded scheduler with MCell version.
The toy_scheduler.c program serves to compare the C++ version
with an original MCell version via a similar interface. This
was originally accomplished by duplicating the original code
and adding an internal user interface. That made it difficult
to ensure that the original MCell version hadn't been modified.
This commit keeps the original code separate to make it clear
what has been modified. Both the sched_util.c and .h files
should be identical to the versions in the MCell repository.
To help keep them identical, they have been saved with read
only permissions. These may or may not be preserved by git.

Note that valgrind's memcheck finds no memory leaks in MCell
when run with a relatively simple MDL model. However, it does
report memory leaks when the scheduler is run with the user
interface provided in the "toy_scheduler" (both C and C++).