Added test_1 (with Oliver's naming convention) and libMCellDM.py
The test_1_unbounded_diffusion.py file follows Oliver's naming
convention (to make it easier to compare with similar alternatives).
It is a fairly traditional version taken from the December 2nd README.md
file.

The new libMCellDM.py builds on a concept that evolved from:

  - Data model only version (also from December 2nd README.dm page)
  - Discussions with Tom about using the __dict__ attributes of a class
  - GoodCode web site: https://goodcode.io/articles/python-dict-object/

This example starts with a data model and builds an object tree that can
be referenced as if they were objects from classes. The idea here is
that different elements of the data model can become different
subclasses of this base class - with each having different functions.
The subclassing and function addition hasn't been done yet.

Since the "traditional" approaches appear to already be well
represented, this is mostly an experiment to see how a different API
approach would end up looking.