Added "path" code to identify specific sections of the data model.
This commit is preparation for subclassing "DataModelObject" to contain
the appropriate functions ("methods") for each type of MCell object.
This may be done dynamically in the same way that attributes are being
added in the present code.

The current code builds all objects of class DataModelObject. However,
it should be possible to choose a more specific class based on the type
of data it contains. Since individual item names (like "x") may be
reused in different parts of the data model (and are therefore not
unique), the entire path from "mcell" down to each element should be
considered when making class assignments. This commit provides that
functionality by tracking the "path" from the mcell root down to each
object in the data model. It matches the path at each point in the
process against the known paths which should eventually map to the
specialized subclasses of DataModelObject. When a known subclass is
detected, the new object can either be created of the proper class or
the specialized methods can be added dynamically.

In the long run, the detection of each object would make a call to
libMCell to add the "live" functionality to the static data model.