Began converting some of the data model objects into specific classes.
When the data model is read, unknown items are created as type
DataModelItem. Items that are recognized (like mcell.model_objects) will
be created as their specific type but containing all of the keys found
in the data model. This allows dynamically created objects to become
specific classes with specific functionality.

This version has only implemented specific classes for:

  Species
  ModelObjects
  GeometryObjects

All other data model items that are recognized by their data model path
(such as "mcell.define_reactions") are assigned to the "DataModelStub"
class. Any that are not recognized by their data model path remain as
DataModelItem.