Changed the logic to initialize in the order items are in path_obj_map
The previous version was driven by a dictionary containing each item to
be initialized from the data model ( parameters, initialization, ...).
However, because they were in a dictionary, their keys were not ordered
and the initialization was somewhat random.

This commit replaces the dictionary with a tuple of tuples. The incoming
data model dictionary keys are then sorted to match this ordering before
the items are constructed. Note that this sorting may not be efficient.