Implemented actual classes for the remaining Data Model groups.
Each class inherits all of the data model values appropriate for that
class through the DataModelItem super class.

Also added a "print(self)" function to each class to print out a
reasonable representation of its data. While each print function is
helpful, the main purpose for writing them was to demonstrate how to
access the items in each class.