Add example 7 (and rxn directionality to 6).
Example 7 deals with reactions at a surface, which meant I had to go
back and add directionality to pymcell.py and example 6. I suppose there
are a couple ways to handle this when you're dealing with only volume
molecules (which don't care about directionality). We could just ignore
the directionality (that's what I'm doing now) or we could do it the
duck-typing way and have it be based off of what is actually passed in
(i.e. w/ orientations for anything involving a surface and w/o
orientations for anything that only has volume molecules).