Mostly Finished Generalized Geometry in test case 5
A better name might be "Callback Geometry" since the idea is that an
object might provide "callback" routines to be used by libMCell when
checking an object for collisions or other geometry-related reasons.

In this simple example, the "my_box" class inherits from MCellGeometry
and over-rides these two methods:

  check_inside ( point )
  point_and_normal_of_first_crossing ( from_point, to_point )

It is imagined that libMCell might call these two functions (and
possibly others) as it does its own processing.

This interface points out that some geometrical objects need not have
any points or faces at all (much like Blender's Meta Ball objects).
These geometry items simply need to be able to answer the questions that
MCell needs to have answered. For example, "Is this location inside your
boundaries?" and "What is the location and normal of the first surface
to be hit when moving in a straight line from here to there?". Some
kinds of "virtual geometry" objects can answer those questions without
keeping any points or faces ... and provide a more accurate answer.

If such an interface is provided in the API, then our "normal" triangle
processing might be implemented to fit that interface. In other words,
it simply provides the functions that the interface requires. This would
be a very clean separation of functionality while giving the end user an
easy way to implement arbitrary object representations (quads, "metas",
"neurons", etc). It also makes it clear how the user could implement
dynamic geometry. After all, the geometry system just needs to answer
the questions based on the current state of the geometry when asked. If
we add additional "force functions" we might get a very natural kind of
dynamic geometry.