Finished up the time-slotted version started before slotted discussion.
The group held a discussion about whether the events should be stored:

  1) In a 1D list sorted by time and other attributes
  2) In a 2D list sorted by time in D1 and other attributes in D2

The discussion concluded with option (1).

The earlier implementations (previous commits) had not yet implemented
the unique time implementation of option (2), but the work had been
partially completed in the working directory. This commit contains the
finished version of option (2) for completeness.