Moved the subclasses directly below the base class for comparison.
The subclasses were moved closer to the base class.

Additional description of this approach (left out of previous commits):

The original "framework_zero" had one execute function to be called when
the scheduled time arrived. That is the simplest mechanism for attaching
any executable code to an event. It is a static class-based approach and
it can be subclassed for a variety of specialized functionality.

One such specialization is support for a dynamic list of callback
functions. The previous version (and this one) implement that
specialization as a subclass of the original functionality which uses
the execute function to call a list of callbacks. This can be seen in
the "schedulable_object_with_callback_list" class which uses the execute
function to loop through all the callbacks in its list.

Note that calling all the functions in a callback list is just ONE such
specialization. Another specialization might call each callback until
one of them returns a "handled" code. This, for example, is one way to
delegate an event through a list of handlers until one of them decides
that it was actually able to handle the event. Once one of the handlers
in the list has handled the event, it's no longer necessary (and likely
an error) to continue to pass the event to additional handlers in the
callback list.

Another specialization would be to have prioritized callbacks where each
callback has a priority value set. Only callbacks with priorities above
the current operating priority would actually be called for that event.

Another specialization would be to include a list of callback *objects*
rather than a list of callback *functions*. A callback object would
contain data as well as its own functions ("execute" or "callback").
This is a relatively powerful approach because it wraps data and
functionality. To get this approach with the callback list base class
would require subclassing the base class and adding a callback *object*
list and completely ignoring the callback *function* list. Using the
original "schedulable_object" class there is no "callback_list" to be
ignored ... since that base class has no data items at all.

The point is that by folding the "callback list" functionality into the
base class, these other options would have to be implemented "around"
the unused callback list. It's generally not a good design to have
subclasses that are forced to ignore data items in the base class,
especially if there's another way to do it.