Added another re-write of scheduler and framework from scratch.
This version was written at home on Saturday. It separates motion
(diffusion) from other aspects of a molecule by containing a point
object within the molecule. That allows different molecule reaction
types to share the same molecule motion types.

The current version is based on a "point" class, but this may be too
specific a term. Better terms might be "location" or "position" which
may include other aspects of location (rotation, for example), and may
even be more generalized ("inside object X").