Updated to use the term "location" rather than "point".
As mentioned in the previous commit, the notion of "location" may
include other represenations than just a point. This change uses the
term "location".

Note also that the following files implement a 2D version. This was done
to make a simpler simulation for checking out linkages between code
elements (scheduler, callbacks, molecules, reactions, motion, display).
These are the current 2D files:

  gui.py
  location.py
  molecules.py
  scheduler.py
  sim_2d.py
  test_scheduler.py

The program can be run from the command line with:   "python gui.py"
This will bring up a GTK window with a few buttons for stepping printing
and resetting. The molecules in this simulation use 4 different motion
classes:

  point_location - no motion
  brownian_point - random motion (not currently correct)
  newtonian_point - constant velocity

  spiral_point - moves in an outward spiral starting at the origin

The first 3 classes are in the location class itself and may be
generally useful. The spiral_point class was created in the "user"
program "sim_2d.py" and is primarily a demonstration of how a user
might subclass a "point_location" to give it custom behavior.