Added some checks to keep from advancing the buffered frames at end.
The GUI code didn't have any way to know that no more events were
scheduled, so it would continue to buffer unchanged frames. These
additional frames would then need to be viewed when going back (which
appeared as if nothing was happening). This minor change stops buffering
frames when the scheduler is empty.