Added quad_tree.py from Algorithms/collision_detection
This python program borrows from a version of backpropagation in:

  https://gitlab.snl.salk.edu/NeuralNetworkCoding/AssortedPrograms

It was modified to show a quad tree of points. The program follows
mouse motion events and highlights those points in the quad tree
cell under the mouse pointer position. This program is an exploration
of an algorithm that might be helpful in libMCell.