Switched from string keys to 2-tuple of integer indicies.
This was done mostly for dumping the SpatialHash in a sorted order (the
tuples of integers sort more properly than strings). It's not clear
whether the hash of a 2-tuple is faster to compute than a hash of a
joined string. Speed should dictate the final choice, but sortability is
helpful during debugging.