Added diffusion / run capability. Fixed recursion depth problem.
Notes for current QuadTree algorithm:

When points are allowed to leave the original box, they are mapped to
the edges and corners. If points diffuse naturally, they will fan out
to cover an infinite space. The region of space containing the original
box and the space directly above, below, left and right will eventually
be small (like points and lines) compared to the entire area of the
diffusion. These large area spaces map to the corners of the original box,
and those corners will continue to subdivide. Eventually, their recursion
depth will overflow the max recursion depth of the process if not checked
by some mechanism.

Here's what it looks like where the original box is represented by the
"#" character in the center:

                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
  =============================#===============================
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H
                               H

There will be a very small fraction of the points remaining in the
original box. There will also be a small fraction of the points that
fall in the two "lines". They will map to the edges. The great majority
of points will be in the 4 large quadrant areas which map to the four
corners. As a result, those 4 corners will eventually contain all of
the points. If each subdivision can only have a small number of points
(less than "max_objects"), then they will have to split many many times
to hold all their points. This splitting creates increased depth which
exceeds the recursion depth.

Note that this is only a problem because the algorithm continues to
subdivide boxes within the original constrained area rather than expanding
its area. It is compounded by a small "max_objects" per box. The small
"max_objects" forces more depth than would otherwise be required.

For a reaction/diffusion system, there should be a limit on the
subdivision that reflects the mean distance a molecule might travel in
a single time step. Subdividing any smaller than that creates more work
when finding collisions. So the smallest size should be on the order of
the distance of reaction partners in a single time step.