Tried to clean up some of the problems in zoom_pan_window.py
It is likely that zoom_pan_window.py was being re-worked to better match
the Java version, but hadn't been completed (back in June?). It's likely
that the matching Java source code was being pasted in and then modified
to be Python (preserving the logic that works in the Java version).

It appears that the process was not complete, and these few edits were
added to try to get it working. Since that may require more work than
expected (right now), it's being left in the current incomplete state.